pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "pbr>=7.0.3",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.1",
]
